add init script and misc utilities
Adds an init script that launches UxAS as the `uxas` user upon
startup. It will look for the UxAS executable at `/home/uxas/uxas`,
will pass it `/home/uxas/cfg.xml` for the `-cfgPath` option, and runs
in the `/home/uxas/run` directory.

Also adds `zip`, and compiles BusyBox's `tar` with support for the
`-z` option to create gzip archives.